src/test/kotlin/org/carth/aoc21/day20/Day20Test.kt
package org.carth.aoc21.day20

import org.carth.aoc21.common.Resources.inputAsListOfString
import org.carth.aoc21.common.Resources.testInputAsListOfString
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.MethodOrderer
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Order
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.TestMethodOrder

private const val DAY = "day20"

class Day20Tests {

    @Nested
    @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
    inner class Day20PartOneTest {

        @Test
        @Order(1)
        fun `solve part one of sample`() {
            // Act
            val answer = Day20(testInputAsListOfString(DAY)).solvePartOne()

            // Assert
            assertEquals(35, answer)
        }

        @Test
        @Order(2)
        fun `solve part one`() {
            // Act
            val answer = Day20(inputAsListOfString(DAY)).solvePartOne()

            // Assert
            assertEquals(5597, answer)
        }
    }

    @Nested
    @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
    inner class Day20PartTwoTest {

        @Test
        @Order(1)
        fun `solve part two of sample`() {
            // Act
            val answer = Day20(testInputAsListOfString(DAY)).solvePartTwo()

            // Assert
            assertEquals(3351, answer)
        }

        @Test
        @Order(2)
        fun `solve part two`() {
            // Act
            val answer = Day20(inputAsListOfString(DAY)).solvePartTwo()
            // Assert
            assertEquals(18723, answer)
        }
    }
}